+++ docs/CHANGE_LOG.md
@@ -4,6 +4,8 @@
 
 #### **Start Options:**
 - Added a start option "offset" (defaults to 0), which states the additional offset to use for the dialogs position.
+- Added a start option "showSearch" (defaults to true), which states if the search bar should be shown.
+- Added a start option "showSearchPropertyCount" (defaults to 15), which states the minimum number of properties that should be available for the search bar to be visible.
 
 #### **Configuration Options:**
 - BREAKING: Renamed the configuration option "dialogDisplayDelay" to "dialogShowDelay".